+++ projects/angular-material-extended/src/lib/matx-autocomplete/matx-autocomplete-template.directive.spec.ts
@@ -1,0 +1,8 @@
+import { MatxAutocompleteTemplateDirective } from './matx-autocomplete-template.directive';
+
+describe('MatxAutocompleteTemplateDirective', () => {
+  it('should create an instance', () => {
+    const directive = new MatxAutocompleteTemplateDirective();
+    expect(directive).toBeTruthy();
+  });
+});