+++ projects/angular-material-extended/src/lib/matx-autocomplete/matx-autocomplete-template.directive.ts
@@ -1,0 +1,4 @@
+import { Directive, Input } from '@angular/core';
+
+@Directive({selector: '[matxAutocompleteTemplate]'})
+export class MatxAutocompleteTemplateDirective {}