+++ projects/angular-material-extended/src/lib/matx-prompt/matx-prompt.component.ts
@@ -42,4 +42,8 @@ export class MatxPromptComponent implements OnInit {
       this.dialogRef.close();
     }
   }
+
+  withForm(value, form: FormGroup) {
+    return typeof value === 'function' ? value(form) : value;
+  }
 }