+++ projects/angular-material-extended/src/public_api.ts
@@ -3,22 +3,32 @@
  */
 
 export * from './lib/matx.module';
+
 export * from './lib/matx-input/matx-input.module';
 export * from './lib/matx-input/matx-input.component';
+
 export * from './lib/matx-errors/matx-errors.module';
 export * from './lib/matx-errors/matx-errors.component';
+
 export * from './lib/matx-autocomplete/matx-autocomplete.module';
 export * from './lib/matx-autocomplete/matx-autocomplete.component';
+
 export * from './lib/matx-menu-button/matx-menu-button.component';
+
 export * from './lib/controllers/matx-sidenav-menu.controller';
+
 export * from './lib/matx-select/matx-select.module';
+
 export * from './lib/matx-select/matx-select.component';
+
 export * from './lib/matx-prompt/matx-prompt.module';
 export * from './lib/matx-prompt/matx-prompt.component';
 export * from './lib/matx-prompt/matx-prompt.controller';
+
 export * from './lib/matx-loading/matx-loading.module';
 export * from './lib/matx-loading/matx-loading.component';
 export * from './lib/matx-loading/matx-loading.controller';
+
 export * from './lib/matx-datepicker/matx-datepicker.component';
 export * from './lib/matx-datepicker/matx-datepicker.module';